+++ src/main/kotlin/com/github/blarc/ai/commits/intellij/plugin/notifications/Notification.kt
@@ -40,6 +40,7 @@ data class Notification(
         )
 
         fun emptyDiff() = Notification(DEFAULT_TITLE, message = message("notifications.empty-diff"))
+        fun promptTooLarge() = Notification(DEFAULT_TITLE, message = message("notifications.prompt-too-large"))
         fun unsuccessfulRequest(message: String) = Notification(message = message("notifications.unsuccessful-request", message))
 
     }